# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Locker Network API",
    "version": "1.0.0",
    "description": "🔐 Locker Network - Secure smart locker management system",
    "contact": {
      "name": "API Support",
      "email": "support@lockernetwork.com"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3555",
      "description": "Development server"
    },
    {
      "url": "https://api.lockernetwork.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT access token"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "httpOnly cookie with refresh token"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": ["USER", "OPERATOR", "ADMIN"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SignupRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "example": "SecurePass123!",
            "description": "Must contain uppercase, lowercase, number, and special character"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "SecurePass123!"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token (use in Authorization header)"
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "LockerBox": {
        "type": "object",
        "properties": {
          "lockerBoxId": {
            "type": "string",
            "format": "uuid"
          },
          "stationId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "example": "A-01"
          },
          "size": {
            "type": "string",
            "enum": ["S", "M", "L"]
          },
          "status": {
            "type": "string",
            "enum": ["AVAILABLE", "RESERVED", "OCCUPIED", "FAULTY", "EXPIRED"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LockerStation": {
        "type": "object",
        "properties": {
          "stationId": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "type": "string",
            "example": "Tel Aviv"
          },
          "address": {
            "type": "string",
            "example": "Dizengoff 10"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE", "MAINTENANCE"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "lockerBoxId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "ACTIVE", "ENDED", "FAILED", "CANCELLED", "EXPIRED"]
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "expectedEndTime": {
            "type": "string",
            "format": "date-time"
          },
          "totalPrice": {
            "type": "number",
            "format": "decimal"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "example": "USER_LOGIN"
          },
          "entityType": {
            "type": "string",
            "example": "User"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication endpoints"
    },
    {
      "name": "Users",
      "description": "User profile endpoints"
    },
    {
      "name": "Lockers",
      "description": "Locker management"
    },
    {
      "name": "Stations",
      "description": "Station management"
    },
    {
      "name": "Bookings",
      "description": "Booking/Reservation endpoints"
    },
    {
      "name": "Audit",
      "description": "Audit logs"
    }
  ],
  "paths": {
    "/api/v1/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new user",
        "description": "Create a new user account. Password must contain uppercase, lowercase, number, and special character (min 12 chars).",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully registered",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "example": "refreshToken=...; Path=/; HttpOnly; Secure; SameSite=Strict; Max-Age=604800"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request or user already exists"
          },
          "429": {
            "description": "Too many signup attempts"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate user",
        "description": "Login with email and password",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many login attempts"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "description": "Get a new access token using refresh token from cookie",
        "operationId": "refreshToken",
        "responses": {
          "200": {
            "description": "New access token generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing refresh token"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth", "Users"],
        "summary": "Get current user profile",
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout user",
        "operationId": "logoutUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}